// frontend/src/app/core/services/api/model/qualificationTierDto.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Data Transfer Object for qualification tier
 */
export interface QualificationTierDto { 
    uuid?: string;
    qualificationSchemeId?: string;
    boatCountMin?: number;
    boatCountMax?: number;
    heatCount?: number;
    semiFinalCount?: number;
    finalBCount?: number;
    finalACount?: number;
    createdAt?: string;
    modifiedAt?: string;
}