// frontend/src/app/core/services/api/model/qualificationProgressionDto.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Data Transfer Object for qualification progression
 */
export interface QualificationProgressionDto { 
    uuid?: string;
    qualificationTierId?: string;
    sourceEvent?: string;
    destEvent?: string;
    qualifyByPosition?: number;
    qualifyByTime?: number;
    createdAt?: string;
    modifiedAt?: string;
}