// frontend/src/app/core/services/api/model/qualificationProgressionRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request payload for creating or updating a qualification progression
 */
export interface QualificationProgressionRequest { 
    qualificationTierId: string;
    sourceEvent: string;
    destEvent: string;
    qualifyByPosition: number;
    qualifyByTime: number;
}